add acl_fiber_kill for fiber lib; add listen callback for master template.

+++ lib_fiber/samples/chat/reader/main.cpp
@@ -14,7 +14,7 @@ static int __rw_timeout = 0;
 static int __max_client = 10;
 static int __cur_client = 10;
 static struct timeval __begin;
-static int __total_read = 0;
+static long long __total_read = 0;
 
 static bool client_login(user_client* uc)
 {
@@ -112,7 +112,7 @@ static void fiber_client(const char* addr, const char* user, int max_loop)
 		struct timeval end;
 		gettimeofday(&end, NULL);
 		double spent = stamp_sub(&end, &__begin);
-		printf("---Total read: %d, spent: %.2f, speed: %.2f---\r\n",
+		printf("---Total read: %lld, spent: %.2f, speed: %.2f---\r\n",
 			__total_read, spent,
 			(1000 * __total_read) / (spent < 1 ? 1 : spent));
